Move adapter classes into separate module
