+++ bridges/ironSorce/adapter/.gitignore
@@ -1,0 +1,1 @@
+/build
