+++ bridges/ironSorce/adapter/build.gradle
@@ -1,17 +1,17 @@
 plugins {
-    id 'com.android.application'
+    id 'com.android.library'
+    id 'maven-publish'
 }
 
 android {
+    namespace 'com.ironsource.adapters.custom'
+    compileSdk 34
+
     defaultConfig {
-        //applicationId "com.loopme.ironsource_mediation_sample"
-        minSdkVersion 24
-        compileSdk 34
-        targetSdkVersion 34
-        versionCode 1
-        versionName "1.0"
+        minSdk 24
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles "consumer-rules.pro"
     }
 
     buildTypes {
@@ -25,7 +25,27 @@ android {
         sourceCompatibility JavaVersion.VERSION_17
         targetCompatibility JavaVersion.VERSION_17
     }
-    namespace 'com.loopme.ironsource_mediation_sample'
+
+    publishing {
+        publishing {
+            singleVariant("release") {
+                withSourcesJar()
+                withJavadocJar()
+            }
+        }
+    }
+}
+
+afterEvaluate {
+    publishing {
+        publications {
+            release(MavenPublication) {
+                artifactId = 'ironsource-adapter'
+
+                from components.release
+            }
+        }
+    }
 }
 
 repositories {
@@ -36,14 +56,11 @@ repositories {
 }
 
 dependencies {
-
     implementation 'androidx.appcompat:appcompat:1.6.1'
-    implementation 'com.google.android.material:material:1.10.0'
-    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
-    implementation 'com.ironsource.sdk:mediationsdk:7.7.0'
-    implementation project(path: ':loopme-sdk')
-    implementation project(path: ':bridges:ironSorce:adapter')
+    implementation 'com.google.android.material:material:1.11.0'
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
+    implementation 'com.ironsource.sdk:mediationsdk:7.7.0'
+    implementation project(path: ':loopme-sdk')
 }
