+++ bridges/ironSorce/adapter/src/main/java/com/ironsource/adapters/loopme/LoopmeCustomAdapter.java
@@ -1,4 +1,4 @@
-package com.ironsource.adapters.custom.loopme;
+package com.ironsource.adapters.loopme;
 
 import android.app.Activity;
 import android.content.Context;
