+++ bridges/ironSorce/adapter/src/main/java/com/ironsource/adapters/loopme/provider/IBannerContainerProvider.java
@@ -1,10 +1,11 @@
-package com.ironsource.adapters.custom.loopme.provider;
+package com.ironsource.adapters.loopme.provider;
 
+import com.ironsource.adapters.loopme.LoopmeCustomBanner;
 import com.ironsource.mediationsdk.IronSourceBannerLayout;
 
 /**
  * The activity hosting the banner view should implement this interface in order to provide
- * {@link com.ironsource.adapters.custom.loopme.LoopmeCustomBanner} with a reference to the
+ * {@link LoopmeCustomBanner} with a reference to the
  * IronSourceBannerLayout view in order to correctly implement mediation.
  */
 public interface IBannerContainerProvider {
