+++ bridges/ironSorce/app/src/main/java/com/loopme/ironsource_mediation_sample/RewardedVideoActivity.java
@@ -3,9 +3,8 @@ package com.loopme.ironsource_mediation_sample;
 import android.app.Activity;
 import android.os.Bundle;
 import android.view.View;
-import android.widget.EditText;
 import android.widget.Toast;
-import com.ironsource.adapters.custom.loopme.LoopmeCustomAdapter;
+import com.ironsource.adapters.loopme.LoopmeCustomAdapter;
 import com.ironsource.mediationsdk.IronSource;
 import com.ironsource.mediationsdk.adunit.adapter.utility.AdInfo;
 import com.ironsource.mediationsdk.logger.IronSourceError;
