+++ bridges/ironSorce/settings.gradle
@@ -1,2 +1,3 @@
 include ':app'
 include ':loopme-sdk'
+include ':adapter'
