+++ settings.gradle
@@ -5,3 +5,4 @@ include ':sample-banner'
 //include ':bridges:mopub:mopub-mediation-sample:app'
 //include ':bridges:appLovin:appLovin-mediation-sample:app'
 include ':bridges:ironSorce:app'
+include ':bridges:ironSorce:adapter'
